EmbeddedPkg: FastbootUsb: use ep2 as OUT endpoint
It's less generic to use ep1 as both IN and OUT endpoint, because some
platforms have data endpoints that work only in single direction.  For
example of Poplar SoC, the ep1 works as IN, and ep2 as OUT endpoint
only.

Before a more generic implementation is available, as a stop-gap
solution, let's change to use ep2 as OUT endpoint, since most of
platforms like Hikey have both ep1 and ep2 work bidirectionally.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>